Returning just a list of doubles for letter frequencies
For the ChiSquaredAnalaysis we need the data points to be in the same order, which we can't really ensure when using the current dictionary structure. Now on loading the JSON, we order the available letters. We then use these ordered available letters when getting the frequencies for each language, and we also use this same order when getting frequencies for an input string to ensure the data points are always in the same index for the same character,